backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.3
pdfminer.six==20231228
sentence-transformers==3.0.1

# Force CPU-only PyTorch (no CUDA / GPU deps)
torch==2.2.2+cpu
torchvision==0.17.2+cpu
torchaudio==2.2.2+cpu
--extra-index-url https://download.pytorch.org/whl/cpu

faiss-cpu==1.8.0
google-generativeai==0.7.2
azure-cognitiveservices-speech==1.38.0
python-multipart==0.0.9
pydantic==2.8.2
python-dotenv==1.0.0
